+++ src/clj/r99c/routes/home.clj
@@ -10,13 +10,25 @@
 (defn home-page [request]
   (layout/render request "home.html" {:docs (-> "docs/docs.md" io/resource slurp)}))
 
+(defn problems-page [request]
+  (layout/render request "problems.html" {:problems (db/problems)}))
 
+(defn answer-page [request]
+  (let [num (get-in request [:path-params :num])
+        problem (db/get-problem {:num (Integer/parseInt num)})]
+    (layout/render request "answer-form.html" {:problem problem})))
+
+(defn create-answer! [request]
+  (layout/render request "request.html" {:request request}))
 
 (defn home-routes []
   [""
    {:middleware [middleware/auth
                  middleware/wrap-csrf
                  middleware/wrap-formats]}
-   ["/" {:get home-page}]])
-   
+   ["/" {:get home-page}]
+   ["/problems" {:get problems-page}]
+   ["/answer/:num" {:get  answer-page
+                    :post create-answer!}]])
+