+++ CHANGELOG.md
@@ -16,11 +16,14 @@ int func_test(void) {
 * answer をボタンに。button is-primary is-small でもやや大きすぎ、ブサイク。
 * get-answer :login が複数返ることもある。
 * login/regisger の説明書き
+* status problems に色つけ
 
-## 0.3.3-SNAPSHOT
+
+## 0.3.3  - 2021-10-10
 ### Added
 * problems ... defined table and a route /problems
 * answers ... defined table and a route /answer:num
+* /status
 
 ## 0.3.2 - 2021-10-09
 ### Added